test: isolate recorder reference video flow
